append subscriber log
